requirements.txt
# Core
torch==2.3.1
torchvision==0.18.1
numpy==1.26.4
scipy==1.13.1
pillow==10.4.0
pandas==2.2.2
tqdm==4.66.4

# Imaging & I/O
scikit-image==0.24.0
opencv-python==4.10.0.84
matplotlib==3.9.0

# Metrics / matching
lapx==0.5.9           # fast Hungarian; fallback is scipy.optimize.linear_sum_assignment

# Optional utilities
einops==0.8.0

